requirements.txt
# Web Framework
fastapi
python-multipart
uvicorn

# --- Core Functionality ---

# Database (Asynchronous MongoDB Driver)
motor

# Data Validation & Settings
pydantic
pydantic-settings
email-validator
python-dotenv

# --- AI & PDF ---

# Anthropic Claude API Client
anthropic

# PDF Text Extraction
pypdf

# --- Authentication & Security ---

# Password Hashing
passlib[bcrypt]
bcrypt==3.2.0

# JWT Handling
python-jose[cryptography]

# --- Job Discovery & Automation ---

# Asynchronous HTTP Client for Scraping
httpx
playwright

# HTML Parsing
beautifulsoup4

# Background Task Scheduling
apscheduler

# --- Rate Limiting ---
redis
fastapi-limiter

# --- JobSpy Multi-Site Scraper ---
python-jobspy
lxml
